+++ examples/shared/src/examples/e2e/map-view/AndroidViewTexture.tsx
@@ -1,0 +1,5 @@
+import { MapView } from "@maplibre/maplibre-react-native";
+
+export function AndroidViewTexture() {
+  return <MapView testID="map-view" androidView="texture" />;
+}