+++ examples/shared/src/examples/e2e/sources/shape-source/GetClusterLeaves.tsx
@@ -62,7 +62,7 @@ export function GetClusterLeaves() {
             id="clusters"
             filter={["has", "point_count"]}
             style={{
-              circleRadius: 20,
+              circleRadius: 80,
               circleColor: colors.blue,
             }}
           />