+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+# [11.0.0-alpha.30](https://github.com/maplibre/maplibre-react-native/compare/v11.0.0-alpha.29...v11.0.0-alpha.30) (2026-01-17)
+
+
+### Bug Fixes
+
+* **MapView:** use codegen for AndroidTextureMapView ([#1210](https://github.com/maplibre/maplibre-react-native/issues/1210)) ([848d3b3](https://github.com/maplibre/maplibre-react-native/commit/848d3b3caf65e58e0342ecc7899fba3c82e4be01))
+
+# [11.0.0-alpha.29](https://github.com/maplibre/maplibre-react-native/compare/v11.0.0-alpha.28...v11.0.0-alpha.29) (2026-01-16)
+
+
+### Bug Fixes
+
+* **MapView:** onPress bubbles from Sources ([#1211](https://github.com/maplibre/maplibre-react-native/issues/1211)) ([d7dee4b](https://github.com/maplibre/maplibre-react-native/commit/d7dee4bfe1d372ad49d4872cf5c901ccfce6541e))
+
 # [11.0.0-alpha.28](https://github.com/maplibre/maplibre-react-native/compare/v11.0.0-alpha.27...v11.0.0-alpha.28) (2026-01-08)